made changes in circleci config file and refactor the project structure
