+++ src/test/java/utils/CustomLoggingListener.java
@@ -1,4 +1,4 @@
-package listener;
+package utils;
 
 import org.openqa.selenium.*;
 import org.openqa.selenium.support.events.AbstractWebDriverEventListener;
